Refactor Handler interface: Merge Start() into Wrap()
Having a separate Start() method executed in a goroutine led to initialization
race conditions where sometimes (?) the Start method was not executed before
value initialized by it were accessed in the Wrap() method.

We can prevent this by combining both methods and rather returning a task to be
executed a a goroutine from wrap if there is a need for it.